refactor(contract): move the foreign-tx request path into api/foreign_chain.rs
