+++ crates/contract/src/lib.rs
@@ -34,25 +34,18 @@ pub use crate::pending_requests::MAX_PENDING_REQUEST_FAN_OUT;
 use std::collections::{BTreeMap, BTreeSet};
 
 use crate::{
-    dto_mapping::args_into_verify_foreign_tx_request, errors::Error,
-    foreign_chains_metadata::ForeignChainsMetadata, primitives::ckd::CKDRequest,
+    errors::Error, foreign_chains_metadata::ForeignChainsMetadata, primitives::ckd::CKDRequest,
     storage_keys::StorageKey, tee::tee_state::TeeState, tee::verifier_votes::TeeVerifierVotes,
     update::ProposedUpdates,
 };
 use config::Config;
-use crypto_shared::types::PublicKeyExtended;
-use errors::{InvalidParameters, InvalidState, RespondError, TeeError};
-use near_mpc_contract_interface::deposits::SIGN_DEPOSIT_YOCTONEAR;
-use near_mpc_contract_interface::method_names;
+use errors::InvalidState;
 use near_mpc_contract_interface::types::Ed25519PublicKey;
-use near_mpc_contract_interface::types::{
-    self as dtos, VerifyForeignTransactionRequest, VerifyForeignTransactionRequestArgs,
-    VerifyForeignTransactionResponse,
-};
+use near_mpc_contract_interface::types::{self as dtos, VerifyForeignTransactionRequest};
 
 use dtos::DomainPurpose;
 use near_sdk::{
-    AccountId, CryptoHash, Gas, NearToken, Promise, PromiseError, PromiseOrValue, env, log, near,
+    AccountId, env, log, near,
     store::{IterableMap, Lazy, LookupMap},
 };
 use node_migrations::NodeMigrations;
@@ -88,7 +81,7 @@ pub struct MpcContract {
     /// The verifier contract account trusted for DCAP verification, or [`None`]
     /// until participants vote one in. An [`Attestation::Dstack`](mpc_attestation::attestation::Attestation::Dstack) submission
     /// offloads quote verification to this account; while it is [`None`], such
-    /// submissions are rejected with [`TeeError::VerifierNotConfigured`].
+    /// submissions are rejected with [`TeeError::VerifierNotConfigured`](crate::errors::TeeError::VerifierNotConfigured).
     // TODO(#3639): once participants have voted a verifier in, make this
     // non-optional via a migration that requires it be set.
     tee_verifier_account_id: Option<AccountId>,
@@ -127,143 +120,15 @@ impl SupportedForeignChainsByNode {
     }
 }
 
-impl MpcContract {
-    fn add_verify_foreign_tx_request(
-        &mut self,
-        request: VerifyForeignTransactionRequest,
-        data_id: CryptoHash,
-    ) {
-        pending_requests::push_pending_yield(
-            &mut self.pending_verify_foreign_tx_requests,
-            request,
-            data_id,
-        );
-    }
-}
+impl MpcContract {}
 
 // User contract API
 #[near]
-impl MpcContract {
-    /// Submit a verification + signing request for a foreign chain transaction.
-    /// MPC nodes will verify the transaction on the foreign chain before signing.
-    /// The signed payload is derived from the transaction ID (hash of tx_id).
-    #[handle_result]
-    #[payable]
-    pub fn verify_foreign_transaction(&mut self, request: VerifyForeignTransactionRequestArgs) {
-        log!(
-            "verify_foreign_transaction: predecessor={:?}, request={:?}",
-            env::predecessor_account_id(),
-            request
-        );
-
-        self.check_request_preconditions(
-            request.domain_id,
-            DomainPurpose::ForeignTx,
-            Gas::from_tgas(self.config.sign_call_gas_attachment_requirement_tera_gas),
-            NearToken::from_yoctonear(SIGN_DEPOSIT_YOCTONEAR),
-        );
-
-        let requested_chain = request.request.chain();
-        let supported_chains = self.get_supported_foreign_chains();
-        if !supported_chains.contains(&requested_chain) {
-            env::panic_str(
-                &InvalidParameters::ForeignChainNotSupported {
-                    requested: requested_chain,
-                }
-                .to_string(),
-            );
-        }
-
-        let callback_gas = Gas::from_tgas(
-            self.config
-                .return_signature_and_clean_state_on_success_call_tera_gas,
-        );
-
-        let request = args_into_verify_foreign_tx_request(request);
-        let callback_args = serde_json::to_vec(&(&request,)).unwrap();
-        self.enqueue_yield_request(
-            method_names::RETURN_VERIFY_FOREIGN_TX_AND_CLEAN_STATE_ON_SUCCESS,
-            callback_args,
-            callback_gas,
-            move |this, id| this.add_verify_foreign_tx_request(request, id),
-        );
-    }
-}
+impl MpcContract {}
 
 // Node API
 #[near]
 impl MpcContract {
-    #[handle_result]
-    pub fn respond_verify_foreign_tx(
-        &mut self,
-        request: VerifyForeignTransactionRequest,
-        response: VerifyForeignTransactionResponse,
-    ) -> Result<(), Error> {
-        let signer = Self::assert_caller_is_signer();
-
-        log!(
-            "respond_verify_foreign_tx: signer={}, request={:?}",
-            &signer,
-            &request
-        );
-
-        self.assert_caller_is_attested_participant_and_protocol_active();
-
-        if !self.protocol_state.is_running_or_resharing() {
-            return Err(InvalidState::ProtocolStateNotRunning.into());
-        }
-
-        if !self.accept_requests {
-            return Err(TeeError::TeeValidationFailed.into());
-        }
-
-        if let Some(expected_payload_hash) = &request.expected_payload_hash
-            && &response.payload_hash != expected_payload_hash
-        {
-            return Err(RespondError::UnexpectedPayloadHash.into());
-        }
-
-        let domain = request.domain_id;
-        let public_key = self.public_key_extended(domain.0.into())?;
-
-        let signature_is_valid = match (&response.signature, public_key) {
-            (
-                dtos::SignatureResponse::Secp256k1(signature_response),
-                PublicKeyExtended::Secp256k1 { near_public_key },
-            ) => {
-                let secp_pk = dtos::Secp256k1PublicKey::try_from(&near_public_key)
-                    .expect("Secp256k1 variant always has a secp256k1 key");
-
-                let payload_hash: [u8; 32] = response.payload_hash.0;
-
-                // Check the signature is correct against the root public key
-                near_mpc_signature_verifier::verify_ecdsa_signature(
-                    signature_response,
-                    &payload_hash,
-                    &secp_pk,
-                )
-                .is_ok()
-            }
-            (signature_response, public_key_requested) => {
-                return Err(RespondError::SignatureSchemeMismatch {
-                    mpc_scheme: Box::new(signature_response.clone()),
-                    user_scheme: Box::new(public_key_requested),
-                }
-                .into());
-            }
-        };
-
-        if !signature_is_valid {
-            return Err(RespondError::InvalidSignature.into());
-        }
-
-        pending_requests::resolve_yields_for(
-            &mut self.pending_verify_foreign_tx_requests,
-            &request,
-            serde_json::to_vec(&response).unwrap(),
-        )
-    }
-
     /// Registers the set of foreign chains the calling node supports.
     ///
     /// Must be called directly from the participant's own NEAR account
@@ -494,21 +359,6 @@ impl MpcContract {
 // Contract developer helper API
 #[near]
 impl MpcContract {
-    /// Presence check for a pending foreign-tx verification request, exposed as a
-    /// view call.
-    ///
-    /// See [`Self::get_pending_request`] for the contract: the returned [`YieldIndex`]
-    /// is an arbitrary representative of a fan-out queue, not "the" yield. Only the
-    /// `Some`/`None` distinction is meaningful.
-    pub fn get_pending_verify_foreign_tx_request(
-        &self,
-        request: &VerifyForeignTransactionRequest,
-    ) -> Option<YieldIndex> {
-        self.pending_verify_foreign_tx_requests
-            .get(request)
-            .and_then(|q| q.first().cloned())
-    }
-
     pub fn get_supported_foreign_chains(&self) -> dtos::SupportedForeignChains {
         let active_participant_account_ids = self
             .protocol_state
@@ -568,39 +418,6 @@ impl MpcContract {
     pub fn get_foreign_chains_configs(&self) -> dtos::ForeignChainsConfigs {
         self.foreign_chains.get().snapshot_by_node()
     }
-
-    /// Yield-resume callback for a single queued foreign-tx verification request.
-    ///
-    /// On success, returns the verification response to the original caller. On
-    /// timeout, pops this yield's slot (the head of the FIFO fan-out queue) from the
-    /// pending-request map and fires `fail_on_timeout` to fail the original
-    /// transaction. Sibling yields queued under the same request key remain pending
-    /// and are cleaned up by their own timeouts (or drained together by a subsequent
-    /// `respond_verify_foreign_tx`).
-    #[private]
-    pub fn return_verify_foreign_tx_and_clean_state_on_success(
-        &mut self,
-        request: VerifyForeignTransactionRequest,
-        #[callback_result] response: Result<VerifyForeignTransactionResponse, PromiseError>,
-    ) -> PromiseOrValue<VerifyForeignTransactionResponse> {
-        match response {
-            Ok(response) => PromiseOrValue::Value(response),
-            Err(_) => {
-                pending_requests::pop_oldest_pending_yield(
-                    &mut self.pending_verify_foreign_tx_requests,
-                    &request,
-                );
-                let fail_on_timeout_gas = Gas::from_tgas(self.config.fail_on_timeout_tera_gas);
-                let promise = Promise::new(env::current_account_id()).function_call(
-                    method_names::FAIL_ON_TIMEOUT.to_string(),
-                    vec![],
-                    NearToken::from_near(0),
-                    fail_on_timeout_gas,
-                );
-                near_sdk::PromiseOrValue::Promise(promise.as_return())
-            }
-        }
-    }
 }
 
 #[cfg(not(target_arch = "wasm32"))]
@@ -628,464 +445,20 @@ mod tests {
     use crate::state::key_event::tests::Environment;
     use crate::state::test_utils::{gen_resharing_state, gen_running_state};
     use crate::tee::tee_state::{NodeAttestation, NodeId};
-    use assert_matches::assert_matches;
-    use dtos::ForeignTxSignPayload;
+
     use dtos::{Curve, DomainConfig, DomainId, Protocol, ReconstructionThreshold};
 
-    use k256::{self, Secp256k1, ecdsa::SigningKey, elliptic_curve};
     use mpc_attestation::attestation::{
         MockAttestation as MpcMockAttestation, VerifiedAttestation,
     };
     use near_mpc_bounded_collections::{NonEmptyBTreeMap, NonEmptyBTreeSet};
 
     use near_mpc_contract_interface::types::DestinationNodeInfo;
-    use near_mpc_contract_interface::types::{
-        BitcoinExtractedValue, BitcoinExtractor, BitcoinRpcRequest, ExtractedValue,
-        ForeignTxPayloadVersion, ForeignTxSignPayloadV1,
-    };
+
     use near_sdk::{NearToken, test_utils::VMContextBuilder, testing_env};
     use primitives::key_state::{AttemptId, KeyForDomain, Keyset};
-    use rand::SeedableRng;
+
     use rand::rngs::OsRng;
-
-    use rstest::rstest;
-
-    /// Register the given foreign chains as supported by all active participants.
-    fn register_supported_chains(
-        contract: &mut MpcContract,
-        chains: impl IntoIterator<Item = dtos::ForeignChain>,
-    ) {
-        let foreign_chain_configuration: dtos::ForeignChainConfiguration = chains
-            .into_iter()
-            .map(|foreign_chain| {
-                (
-                    foreign_chain,
-                    NonEmptyBTreeSet::new(dtos::RpcProvider {
-                        rpc_url: "dummy_url.com".to_string(),
-                    }),
-                )
-            })
-            .collect::<BTreeMap<dtos::ForeignChain, NonEmptyBTreeSet<dtos::RpcProvider>>>()
-            .into();
-
-        // pub struct ForeignChainConfiguration(BTreeMap<ForeignChain, NonEmptyBTreeSet<RpcProvider>>);
-
-        let participants: Vec<_> = contract
-            .protocol_state
-            .active_participants()
-            .participants()
-            .iter()
-            .map(|(account_id, _, _)| account_id.clone())
-            .collect();
-        for account_id in participants {
-            let _env = Environment::new(None, Some(account_id), None);
-            contract
-                .register_foreign_chain_config(foreign_chain_configuration.clone())
-                .expect("register should succeed");
-        }
-    }
-
-    #[test]
-    fn verify_foreign_transaction__should_queue_duplicates_from_different_callers() {
-        // Given: two different callers will submit the same foreign-tx verification request.
-        let mut rng = rand::rngs::StdRng::from_seed([42u8; 32]);
-        let (context, mut contract, secret_key) =
-            basic_setup_with_protocol(Protocol::CaitSith, DomainPurpose::ForeignTx, &mut rng);
-        register_supported_chains(&mut contract, [dtos::ForeignChain::Bitcoin]);
-        let SharedSecretKey::Secp256k1(secret_key) = secret_key else {
-            unreachable!();
-        };
-
-        let request_args = VerifyForeignTransactionRequestArgs {
-            domain_id: DomainId::default().0.into(),
-            payload_version: ForeignTxPayloadVersion::V1,
-            expected_payload_hash: None,
-            request: dtos::ForeignChainRpcRequest::Bitcoin(BitcoinRpcRequest {
-                tx_id: [7u8; 32].into(),
-                confirmations: 2.into(),
-                extractors: vec![BitcoinExtractor::BlockHash],
-            }),
-        };
-        let request = args_into_verify_foreign_tx_request(request_args.clone());
-
-        // When: caller alice submits the request.
-        let alice = AccountId::from_str("alice.near").unwrap();
-        testing_env!(
-            VMContextBuilder::new()
-                .signer_account_id(alice.clone())
-                .predecessor_account_id(alice)
-                .current_account_id(context.current_account_id.clone())
-                .attached_deposit(NearToken::from_yoctonear(1))
-                .build()
-        );
-        contract.verify_foreign_transaction(request_args.clone());
-
-        // And: caller bob submits the identical request — a different account would today
-        // be blocked from receiving a response by alice's submission.
-        let bob = AccountId::from_str("bob.near").unwrap();
-        testing_env!(
-            VMContextBuilder::new()
-                .signer_account_id(bob.clone())
-                .predecessor_account_id(bob)
-                .current_account_id(context.current_account_id.clone())
-                .attached_deposit(NearToken::from_yoctonear(1))
-                .build()
-        );
-        contract.verify_foreign_transaction(request_args);
-
-        // Then: both yields are queued under the single (caller-agnostic) request key.
-        assert_eq!(
-            contract
-                .pending_verify_foreign_tx_requests
-                .get(&request)
-                .map(|q| q.len()),
-            Some(2),
-            "duplicate foreign-tx requests from different callers should fan out",
-        );
-
-        // When: a single valid response is delivered.
-        let payload = ForeignTxSignPayload::V1(ForeignTxSignPayloadV1 {
-            request: request.request.clone(),
-            values: vec![ExtractedValue::BitcoinExtractedValue(
-                BitcoinExtractedValue::BlockHash([42u8; 32].into()),
-            )],
-        });
-        let payload_hash_arr = payload.compute_msg_hash().unwrap().0;
-        let secret_key_ec: elliptic_curve::SecretKey<Secp256k1> =
-            elliptic_curve::SecretKey::from_bytes(&secret_key.to_bytes()).unwrap();
-        let signing_key = SigningKey::from_bytes(&secret_key_ec.to_bytes()).unwrap();
-        let (signature, recovery_id) = signing_key
-            .sign_prehash_recoverable(&payload_hash_arr)
-            .unwrap();
-        let response = VerifyForeignTransactionResponse {
-            payload_hash: payload.compute_msg_hash().unwrap(),
-            signature: dtos::SignatureResponse::Secp256k1(
-                dtos::K256Signature::from_ecdsa_recoverable(&signature, recovery_id),
-            ),
-        };
-
-        with_active_participant_and_attested_context(&contract);
-        contract
-            .respond_verify_foreign_tx(request.clone(), response)
-            .expect("respond_verify_foreign_tx should succeed");
-
-        // Then: both queued yields are drained from the single map entry.
-        assert!(
-            contract
-                .pending_verify_foreign_tx_requests
-                .get(&request)
-                .is_none()
-        );
-    }
-
-    #[test]
-    fn respond_verify_foreign_tx__should_succeed_when_response_is_valid_and_request_exists() {
-        // Given
-        let mut rng = rand::rngs::StdRng::from_seed([42u8; 32]);
-        let (context, mut contract, secret_key) =
-            basic_setup_with_protocol(Protocol::CaitSith, DomainPurpose::ForeignTx, &mut rng);
-        register_supported_chains(&mut contract, [dtos::ForeignChain::Bitcoin]);
-        testing_env!(context.clone());
-        let SharedSecretKey::Secp256k1(secret_key) = secret_key else {
-            unreachable!();
-        };
-        let request_args = VerifyForeignTransactionRequestArgs {
-            domain_id: DomainId::default().0.into(),
-            payload_version: ForeignTxPayloadVersion::V1,
-            expected_payload_hash: None,
-            request: dtos::ForeignChainRpcRequest::Bitcoin(BitcoinRpcRequest {
-                tx_id: [7u8; 32].into(),
-                confirmations: 2.into(),
-                extractors: vec![BitcoinExtractor::BlockHash],
-            }),
-        };
-
-        // When
-        let request = args_into_verify_foreign_tx_request(request_args.clone());
-        contract.verify_foreign_transaction(request_args);
-        contract
-            .get_pending_verify_foreign_tx_request(&request)
-            .unwrap();
-        let payload = ForeignTxSignPayload::V1(ForeignTxSignPayloadV1 {
-            request: request.request.clone(),
-            values: vec![ExtractedValue::BitcoinExtractedValue(
-                BitcoinExtractedValue::BlockHash([42u8; 32].into()),
-            )],
-        });
-        let response = sign_foreign_tx_payload(&secret_key, &payload);
-
-        with_active_participant_and_attested_context(&contract);
-
-        // Then
-        match contract.respond_verify_foreign_tx(request.clone(), response.clone()) {
-            Ok(_) => {
-                contract
-                    .return_verify_foreign_tx_and_clean_state_on_success(
-                        request.clone(),
-                        Ok(response),
-                    )
-                    .detach();
-
-                assert!(
-                    contract
-                        .get_pending_verify_foreign_tx_request(&request)
-                        .is_none(),
-                );
-            }
-            Err(_) => panic!("respond_verify_foreign_tx should not fail"),
-        }
-    }
-
-    #[test]
-    fn respond_verify_foreign_tx__should_reject_response_with_unexpected_payload_hash() {
-        // Given
-        let mut rng = rand::rngs::StdRng::from_seed([42u8; 32]);
-        let (context, mut contract, secret_key) =
-            basic_setup_with_protocol(Protocol::CaitSith, DomainPurpose::ForeignTx, &mut rng);
-        register_supported_chains(&mut contract, [dtos::ForeignChain::Bitcoin]);
-        testing_env!(context.clone());
-        let SharedSecretKey::Secp256k1(secret_key) = secret_key else {
-            unreachable!();
-        };
-        let request_args = VerifyForeignTransactionRequestArgs {
-            domain_id: DomainId::default().0.into(),
-            payload_version: ForeignTxPayloadVersion::V1,
-            expected_payload_hash: Some(dtos::Hash256([1u8; 32])),
-            request: dtos::ForeignChainRpcRequest::Bitcoin(BitcoinRpcRequest {
-                tx_id: [7u8; 32].into(),
-                confirmations: 2.into(),
-                extractors: vec![BitcoinExtractor::BlockHash],
-            }),
-        };
-        let request = args_into_verify_foreign_tx_request(request_args.clone());
-        contract.verify_foreign_transaction(request_args);
-        let payload = ForeignTxSignPayload::V1(ForeignTxSignPayloadV1 {
-            request: request.request.clone(),
-            values: vec![ExtractedValue::BitcoinExtractedValue(
-                BitcoinExtractedValue::BlockHash([42u8; 32].into()),
-            )],
-        });
-        let response = sign_foreign_tx_payload(&secret_key, &payload);
-        with_active_participant_and_attested_context(&contract);
-
-        // When
-        let result = contract.respond_verify_foreign_tx(request.clone(), response);
-
-        // Then
-        assert_matches!(
-            result.unwrap_err(),
-            Error::Respond(RespondError::UnexpectedPayloadHash)
-        );
-        assert!(
-            contract
-                .get_pending_verify_foreign_tx_request(&request)
-                .is_some(),
-            "the pending request must remain unresolved",
-        );
-    }
-
-    #[test]
-    fn respond_verify_foreign_tx__should_succeed_when_response_matches_expected_payload_hash() {
-        // Given
-        let mut rng = rand::rngs::StdRng::from_seed([42u8; 32]);
-        let (context, mut contract, secret_key) =
-            basic_setup_with_protocol(Protocol::CaitSith, DomainPurpose::ForeignTx, &mut rng);
-        register_supported_chains(&mut contract, [dtos::ForeignChain::Bitcoin]);
-        testing_env!(context.clone());
-        let SharedSecretKey::Secp256k1(secret_key) = secret_key else {
-            unreachable!();
-        };
-        let rpc_request = dtos::ForeignChainRpcRequest::Bitcoin(BitcoinRpcRequest {
-            tx_id: [7u8; 32].into(),
-            confirmations: 2.into(),
-            extractors: vec![BitcoinExtractor::BlockHash],
-        });
-        let payload = ForeignTxSignPayload::V1(ForeignTxSignPayloadV1 {
-            request: rpc_request.clone(),
-            values: vec![ExtractedValue::BitcoinExtractedValue(
-                BitcoinExtractedValue::BlockHash([42u8; 32].into()),
-            )],
-        });
-        let request_args = VerifyForeignTransactionRequestArgs {
-            domain_id: DomainId::default().0.into(),
-            payload_version: ForeignTxPayloadVersion::V1,
-            expected_payload_hash: Some(payload.compute_msg_hash().unwrap()),
-            request: rpc_request,
-        };
-        let request = args_into_verify_foreign_tx_request(request_args.clone());
-        contract.verify_foreign_transaction(request_args);
-        let response = sign_foreign_tx_payload(&secret_key, &payload);
-        with_active_participant_and_attested_context(&contract);
-
-        // When
-        let result = contract.respond_verify_foreign_tx(request.clone(), response);
-
-        // Then
-        assert!(
-            result.is_ok(),
-            "response matching the expected payload hash must be accepted: {result:?}",
-        );
-    }
-
-    #[test]
-    fn respond_verify_foreign_tx__should_reject_request_with_erased_expected_payload_hash() {
-        // Given
-        let mut rng = rand::rngs::StdRng::from_seed([42u8; 32]);
-        let (context, mut contract, secret_key) =
-            basic_setup_with_protocol(Protocol::CaitSith, DomainPurpose::ForeignTx, &mut rng);
-        register_supported_chains(&mut contract, [dtos::ForeignChain::Bitcoin]);
-        testing_env!(context.clone());
-        let SharedSecretKey::Secp256k1(secret_key) = secret_key else {
-            unreachable!();
-        };
-        let request_args = VerifyForeignTransactionRequestArgs {
-            domain_id: DomainId::default().0.into(),
-            payload_version: ForeignTxPayloadVersion::V1,
-            expected_payload_hash: Some(dtos::Hash256([1u8; 32])),
-            request: dtos::ForeignChainRpcRequest::Bitcoin(BitcoinRpcRequest {
-                tx_id: [7u8; 32].into(),
-                confirmations: 2.into(),
-                extractors: vec![BitcoinExtractor::BlockHash],
-            }),
-        };
-        let request = args_into_verify_foreign_tx_request(request_args.clone());
-        contract.verify_foreign_transaction(request_args);
-        let payload = ForeignTxSignPayload::V1(ForeignTxSignPayloadV1 {
-            request: request.request.clone(),
-            values: vec![ExtractedValue::BitcoinExtractedValue(
-                BitcoinExtractedValue::BlockHash([42u8; 32].into()),
-            )],
-        });
-        let response = sign_foreign_tx_payload(&secret_key, &payload);
-        with_active_participant_and_attested_context(&contract);
-
-        // When
-        let tampered_request = VerifyForeignTransactionRequest {
-            expected_payload_hash: None,
-            ..request.clone()
-        };
-        let result = contract.respond_verify_foreign_tx(tampered_request, response);
-
-        // Then
-        assert_matches!(
-            result.unwrap_err(),
-            Error::InvalidParameters(InvalidParameters::RequestNotFound)
-        );
-        assert!(
-            contract
-                .get_pending_verify_foreign_tx_request(&request)
-                .is_some(),
-            "the pending request must remain unresolved",
-        );
-    }
-
-    fn sign_foreign_tx_payload(
-        secret_key: &k256::Scalar,
-        payload: &ForeignTxSignPayload,
-    ) -> VerifyForeignTransactionResponse {
-        let payload_hash = payload.compute_msg_hash().unwrap();
-        let secret_key_ec: elliptic_curve::SecretKey<Secp256k1> =
-            elliptic_curve::SecretKey::from_bytes(&secret_key.to_bytes()).unwrap();
-        let secret_key = SigningKey::from_bytes(&secret_key_ec.to_bytes()).unwrap();
-        let (signature, recovery_id) = secret_key
-            .sign_prehash_recoverable(&payload_hash.0)
-            .unwrap();
-        let signature = dtos::SignatureResponse::Secp256k1(
-            dtos::K256Signature::from_ecdsa_recoverable(&signature, recovery_id),
-        );
-        VerifyForeignTransactionResponse {
-            payload_hash,
-            signature,
-        }
-    }
-
-    #[test]
-    fn test_verify_foreign_tx_timeout() {
-        // Given
-        let mut rng = rand::rngs::StdRng::from_seed([42u8; 32]);
-        let (context, mut contract, _secret_key) =
-            basic_setup_with_protocol(Protocol::CaitSith, DomainPurpose::ForeignTx, &mut rng);
-        register_supported_chains(&mut contract, [dtos::ForeignChain::Bitcoin]);
-        testing_env!(context.clone());
-        let request_args = VerifyForeignTransactionRequestArgs {
-            domain_id: DomainId::default().0.into(),
-            payload_version: ForeignTxPayloadVersion::V1,
-            expected_payload_hash: None,
-            request: dtos::ForeignChainRpcRequest::Bitcoin(BitcoinRpcRequest {
-                tx_id: [7u8; 32].into(),
-                confirmations: 2.into(),
-                extractors: vec![BitcoinExtractor::BlockHash],
-            }),
-        };
-        let request = args_into_verify_foreign_tx_request(request_args.clone());
-
-        // When
-        contract.verify_foreign_transaction(request_args);
-
-        // Then
-        // assert_matches! requires Debug, which PromiseOrValue doesn't implement
-        assert!(matches!(
-            contract.return_verify_foreign_tx_and_clean_state_on_success(
-                request.clone(),
-                Err(PromiseError::Failed)
-            ),
-            PromiseOrValue::Promise(_)
-        ));
-        assert!(
-            contract
-                .get_pending_verify_foreign_tx_request(&request)
-                .is_none()
-        );
-    }
-
-    #[rstest]
-    #[case(Protocol::CaitSith, DomainPurpose::Sign)]
-    #[case(Protocol::ConfidentialKeyDerivation, DomainPurpose::CKD)]
-    #[should_panic(expected = "this method requires ForeignTx")]
-    fn verify_foreign_tx__should_reject_non_foreign_tx_domain(
-        #[case] protocol: Protocol,
-        #[case] purpose: DomainPurpose,
-    ) {
-        // Given
-        let mut rng = rand::rngs::StdRng::from_seed([42u8; 32]);
-        let (_context, mut contract, _sk) = basic_setup_with_protocol(protocol, purpose, &mut rng);
-
-        // When
-        contract.verify_foreign_transaction(VerifyForeignTransactionRequestArgs {
-            domain_id: DomainId::default().0.into(),
-            payload_version: ForeignTxPayloadVersion::V1,
-            expected_payload_hash: None,
-            request: dtos::ForeignChainRpcRequest::Bitcoin(BitcoinRpcRequest {
-                tx_id: [7u8; 32].into(),
-                confirmations: 2.into(),
-                extractors: vec![BitcoinExtractor::BlockHash],
-            }),
-        });
-    }
-
-    #[test]
-    #[should_panic(expected = "Requested foreign chain, Bitcoin, is not supported.")]
-    fn verify_foreign_tx__should_reject_chain_not_in_policy() {
-        // Given
-        let mut rng = rand::rngs::StdRng::from_seed([42u8; 32]);
-        let (context, mut contract, _sk) =
-            basic_setup_with_protocol(Protocol::CaitSith, DomainPurpose::ForeignTx, &mut rng);
-        // Supported chains has Solana but not Bitcoin
-        register_supported_chains(&mut contract, [dtos::ForeignChain::Solana]);
-        testing_env!(context.clone());
-
-        // When - requesting Bitcoin which is not in the policy
-        contract.verify_foreign_transaction(VerifyForeignTransactionRequestArgs {
-            domain_id: DomainId::default().0.into(),
-            payload_version: ForeignTxPayloadVersion::V1,
-            expected_payload_hash: None,
-            request: dtos::ForeignChainRpcRequest::Bitcoin(BitcoinRpcRequest {
-                tx_id: [7u8; 32].into(),
-                confirmations: 2.into(),
-                extractors: vec![BitcoinExtractor::BlockHash],
-            }),
-        });
-    }
 
     #[test]
     #[should_panic(expected = "Caller must be the signer account")]
